search filter, search, create order fail

+++ do-an-v1/src/main/java/longND/fpt/home/dto/CartDto.java
@@ -20,7 +20,7 @@ public class CartDto {
 	private Long userId;
 	private List<CartItemDto> cartItemDtos;
 	private double totalPrice;
-	private double totalItem;
+	private int totalItem;
 	private double totalDiscountedPrice;
 	private double discounte;
 }
